+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 ### Changed
+**Releases 0.0.26**
+- mail never gets posted to the server. [Issue #20](https://github.com/joegasewicz/react-google-oauth2.0/issues/20)
+- Make get Access token function public. [Issue #23](https://github.com/joegasewicz/react-google-oauth2.0/issues/23)
+- Add custom request headers for dynamic resources [Issue #22](https://github.com/joegasewicz/react-google-oauth2.0/issues/22)
+- Make IOAuthState public [Issue #24](https://github.com/joegasewicz/react-google-oauth2.0/issues/24)
+- Fix incorrect types  [Issue #25](https://github.com/joegasewicz/react-google-oauth2.0/issues/25)
+
 **Releases 0.0.25** - css class selectors are not propagating down to the button component. [Issue #18](https://github.com/joegasewicz/react-google-oauth2.0/issues/18)
 
 **Releases 0.0.24** - Add css class names to button, fixed tests. [Issue #16](https://github.com/joegasewicz/react-google-oauth2.0/issues/16)
@@ -11,4 +18,3 @@
 
 
 ## Unreleased
-- mail never gets posted to the server. [Issue #20](https://github.com/joegasewicz/react-google-oauth2.0/issues/20)